Renamed train_tune.py to train.py + main() function
- deleted train.py and tune.py before merge
- renamed train_tune.py to train.py
- moved code in the main guard into a main() function
