--- project/classifier_2d/mlops/tune.py
@@ -1,210 +1,0 @@
-import configparser
-import logging
-import multiprocessing
-import os
-
-import mlflow
-import optuna
-import pytorch_lightning as pl
-from pytorch_lightning.callbacks import LearningRateMonitor, ModelCheckpoint, EarlyStopping
-from torch.cuda import is_available as cuda_available
-
-from src.data_import_xnat import DataImportXNAT
-import json
-
-from src.datamodule import DataModule
-from src.network import Network
-
-logger = logging.getLogger(__name__)
-
-# Obtain hyperparameters for this trial
-def suggest_hyperparameters(trial):
-
-    dropout = trial.suggest_float("dropout", 0.0, 0.3, step=0.1)
-    lr = trial.suggest_float("lr", 1e-5, 1e-3, log=True)
-    max_lr = trial.suggest_categorical("max_lr",[1e-4,5e-4,1e-3,5e-3])
-    model = trial.suggest_categorical("model", ["convnextv2_tiny.fcmae_ft_in22k_in1k","convnextv2_base.fcmae_ft_in22k_in1k"])
-    batch_size = trial.suggest_categorical("batch_size", [16, 32, 64])
-    pretrained = trial.suggest_categorical("pretrained", [True, False])
-    label_smoothing = trial.categorical("label_smoothing", [0.0, 0.05, 0.1])
-    grad_batches = trial.suggest_int('grad_batches',1,4)
-
-    params = {
-        "dropout": dropout,
-        "lr": lr,
-        "max_lr": max_lr,
-        "model": model,
-        "batch_size": batch_size,
-        "pretrained": pretrained,
-        "label_smoothing": label_smoothing,
-        "grad_batches": grad_batches,
-    }
-    return params
-
-def objective(trial,data,config):
-
-    max_workers = config['system']['max_workers']
-    num_workers = (
-        max_workers
-        if max_workers < multiprocessing.cpu_count()
-        else multiprocessing.cpu_count()
-    )
-
-    best_val_loss = float('Inf')
-    
-    # Start a new mlflow run
-    with mlflow.start_run(nested=True):
-
-        params = suggest_hyperparameters(trial)
-        mlflow.log_params(params)
-  
-        label_dict = json.loads(config['project']['label_dict'])
-
-        # initialise network and datamodule
-        dm = DataModule(
-            data = data,
-            label_dict = label_dict,
-            batch_size = params['batch_size'],
-            validation_fraction = float(config['params']['validation_fraction']),
-            test_fraction = float(config['params']['test_fraction']),
-            num_workers = num_workers,
-            random_seed = int(config['params']['random_seed']),
-            image_size = int(config['params']['image_size'])
-        )
-
-        dm.setup()
-
-        n_classes = len(set([x for x in label_dict.values() if x is not None]))
-        mlflow.log_param('n_classes', n_classes)
-
-        train_class_weights = dm.data_manifest["train"]["class_weights"]
-        validation_class_weights = dm.data_manifest["validation"]["class_weights"]
-        test_class_weights = dm.data_manifest["test"]["class_weights"]
-
-        net = Network(
-            n_classes = n_classes,
-            label_dict = label_dict,
-            model_name = params['model'],
-            pretrained = params['pretrained'],
-            learning_rate = params['lr'],
-            max_lr = params['max_lr'],
-            batch_size = params['batch_size'],
-            dropout = params['dropout'],
-            train_class_weights = train_class_weights,
-            validation_class_weights = validation_class_weights,
-            test_class_weights = test_class_weights,
-            weighted_loss = config.getboolean('params', 'weighted_loss'),
-            loss_fcn = config['params']['loss_fcn'],
-            weight_decay = float(config['params']['weight_decay']),
-            mixup_alpha = float(config['params']['mixup_alpha']),
-            cutmix_alpha = float(config['params']['cutmix_alpha']),
-            mixup_prob = float(config['params']['mixup_prob']),
-            mixup_switch_prob = float(config['params']['mixup_switch_prob']),
-            mixup_mode = config['params']['mixup_mode'],
-            label_smoothing = params['label_smoothing'],
-        )
-
-        # Callbacks
-        callbacks = []
-        callbacks.append(LearningRateMonitor(logging_interval="step"))
-        checkpoint_callback = ModelCheckpoint(
-            save_top_k=1,
-            monitor="val_loss",
-            mode="min",
-            dirpath= f"./checkpoint_trial_{trial.number}/",
-        )
-        callbacks.append(checkpoint_callback)
-
-        early_stopping_callback = EarlyStopping(
-            monitor="val_loss",
-            patience=int(config['params']['patience']),
-            mode="min",
-            verbose=True,
-        )
-        callbacks.append(early_stopping_callback)
-
-        # configure trainer
-        trainer = pl.Trainer(
-            precision="32" if cuda_available() else "16",
-            callbacks=callbacks,
-            devices= 1 if cuda_available() else "auto",
-            accelerator="gpu" if cuda_available() else "auto",
-            log_every_n_steps=1,
-            num_sanity_val_steps=0,
-            max_epochs=int(config['params']['max_epochs']),
-            accumulate_grad_batches= params['grad_batches']
-        )
-
-        trainer.fit(net, dm)
-
-        best_model_path = checkpoint_callback.best_model_path
-        best_model = net.load_from_checkpoint(best_model_path)
-        net.evaluate_best_model(best_model)
-
-        val_loss = checkpoint_callback.best_model_score.item()
-
-        if val_loss <= best_val_loss:
-            best_val_loss = val_loss
-
-    return best_val_loss
-
-def tune(config):
-
-    os.environ["CUDA_VISIBLE_DEVICES"] = config["system"]["cuda_visible_devices"]
-    pl.seed_everything(int(config['system']['random_seed']), workers=True)
-
-    xnat_configuration = {'server': config['xnat']['server'],
-                          'user': config['xnat']['user'],
-                          'password': config['xnat']['password'],
-                          'project': config['xnat']['project'],
-                          'verify': config.getboolean('xnat', 'verify')}
-
-    max_workers = config['system']['max_workers']
-    num_workers = (
-        max_workers
-        if max_workers < multiprocessing.cpu_count()
-        else multiprocessing.cpu_count()
-    )
-
-    importer = DataImportXNAT(
-        xnat_configuration = xnat_configuration,
-        num_workers = num_workers
-        )
-
-    # Import raw data
-    raw_data = importer.import_xnat_data()
-
-    # Download images from XNAT
-    data = importer.xnat_image_download(raw_data)
-    
-    mlflow.pytorch.autolog(log_models=False)
-
-    # Create optuna study (hyperparameter tuning framework)
-    study = optuna.create_study(study_name="project-tune", direction="minimize")
-    study.optimize(lambda trial: objective(trial, data, config), n_trials=int(config['tune']['n_trials']))
-
-    with open(('tune_log.txt'), 'w') as f:
-        f.write("Study statistics: \n")
-        f.write(f"  Number of finished trials: {len(study.trials)}\n")
-        f.write("Best trial:\n")
-
-        trial = study.best_trial
-
-        f.write(f"  Trial number: {trial.number}\n")
-        f.write(f"  Loss (trial value): {trial.value}\n")
-
-        f.write("Params:\n")
-        for key, value in trial.params.items():
-            f.write(f"      {key} = {value}\n")
-    
-    mlflow.log_artifact('tune_log.txt')
-    mlflow.log_artifact('val_transform_failures.csv')
-    mlflow.log_artifact('train_transform_failures.csv')
-    
-if __name__ == '__main__':
-
-    config_path = 'config/config.cfg'
-
-    config = configparser.ConfigParser()
-    config.read(config_path)
-    tune(config)
